requirements-dev.txt
envyaml
pandas
numpy
tensorflow
google-cloud-aiplatform